feat(actions): Fixes
New generic class for actions (FormAction, ItemAction, BulkAction) - ResourceAction. Renaming a method "confirmation" to "isConfirmed" in a trait WithConfirmation
