+++ resources/views/crud/shared/item-action-item.blade.php
@@ -1,4 +1,4 @@
-@if($action->confirmation())
+@if($action->isConfirmed())
     @include('moonshine::crud.shared.confirm-action', ['action' => $action, 'href' => $resource->route('actions.item', $resource->getItem()->getKey(), request()->routeIs('*.query-tag') ? ['index' => $index, 'redirect_back' => 1] : ['index' => $index])])
 @else
     <x-dynamic-component
